requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.0.0
psycopg2==2.9.7
python-environ==0.4.54
sqlparse==0.4.4
typing_extensions==4.7.1
